# fix: reduce slice allocations in route dispatch 