@MattBrittan 2025-07-20
Sorry for the long delay in responding to this, as you say it's a very subtile optimization but, I guess, the router forms part of the hot path so it's worth the slight decrease in readability. Thanks for the PR.